Move spreadsheet creation into commonServicesToolkit
